Rewritten main script
