+++ TypeDB/.all-versions-generator.py
@@ -1,164 +1,88 @@
 import requests
 
-repo = "https://api.github.com/repos/vaticle/typedb/releases"
-filename_all = "all-versions.adoc"
-filename_latest = "latest-version.adoc"
-errors = []
+GITHUB_REPO = "vaticle/typedb"
+API_URL = f"https://api.github.com/repos/{GITHUB_REPO}/releases"
+FILENAME_ALL = "all-versions.adoc"
+FILENAME_LATEST = "latest-version.adoc"
+ERRORS = []
 
 
 def check_url(url):
-    if requests.head(url).status_code < 400:
-        return True
-    else:
-        return False
+    """Check if the URL exists and accessible (status code < 400)."""
+    return requests.head(url).ok
 
 
 def get_versions(url):
-    x = requests.get(url + "?per_page=100")
-    json_data = x.json()
+    """Fetch all versions from the GitHub API and process them."""
     result = []
-    count = 0
-    for json_element in json_data:
-        count += 1
-        if "TypeDB" in json_element["name"]:
-            release = {"version": json_element["tag_name"],
-                       "release_notes": json_element["html_url"],
-                       "win":
-                           {
-                               "url": "",
-                               "hash": "",
-                               "check": ""
-                           },
-                       "lin":
-                           {
-                               "url": "",
-                               "url-arm": "",
-                               "url-x86": "",
-                               "hash": "",
-                               "check": ""
-                           },
-                       "mac":
-                           {
-                               "url": "",
-                               "url-arm": "",
-                               "url-x86": "",
-                               "hash": "",
-                               "check": ""
-                           }
-                       }
-            for asset in json_element["assets"]:
-                # print(asset)
-                if "typedb-all-linux" in asset["name"]:
-                    if "typedb-all-linux-arm64" in asset["name"]:
-                        # print(asset["browser_download_url"])
-                        release["lin"]["url-arm"] = asset["browser_download_url"]
-                        if not check_url(asset["browser_download_url"]):
-                            errors.append(asset["browser_download_url"])
-                            release["lin"]["check"] += "Fail "
-                        else:
-                            release["lin"]["check"] += "PASSED "
-
-                    elif "typedb-all-linux-x86_64" in asset["name"]:
-                        # print(asset["browser_download_url"])
-                        release["lin"]["url-x86"] = asset["browser_download_url"]
-                        if not check_url(asset["browser_download_url"]):
-                            errors.append(asset["browser_download_url"])
-                            release["lin"]["check"] += "Fail "
-                        else:
-                            release["lin"]["check"] += "PASSED "
-
-                    else:
-                        # print(asset["browser_download_url"])
-                        release["lin"]["url"] = asset["browser_download_url"]
-                        if not check_url(asset["browser_download_url"]):
-                            errors.append(asset["browser_download_url"])
-                            release["lin"]["check"] = "Fail"
-                        else:
-                            release["lin"]["check"] = "PASSED"
-
-                elif "typedb-all-windows" in asset["name"]:
-                    # print(asset["browser_download_url"])
-                    if not check_url(asset["browser_download_url"]):
-                        errors.append(asset["browser_download_url"])
-                        release["win"]["check"] = "Fail"
-                    else:
-                        release["win"]["check"] = "PASSED"
-                    release["win"]["url"] = asset["browser_download_url"]
-
-                elif "typedb-all-mac" in asset["name"]:
-
-                    if "typedb-all-mac-arm64" in asset["name"]:
-                        # print(asset["browser_download_url"])
-                        release["mac"]["url-arm"] = asset["browser_download_url"]
-                        if not check_url(asset["browser_download_url"]):
-                            errors.append(asset["browser_download_url"])
-                            release["mac"]["check"] += "Fail "
-                        else:
-                            release["mac"]["check"] += "PASSED "
-
-                    elif "typedb-all-mac-x86_64" in asset["name"]:
-                        # print(asset["browser_download_url"])
-                        release["mac"]["url-x86"] = asset["browser_download_url"]
-                        if not check_url(asset["browser_download_url"]):
-                            errors.append(asset["browser_download_url"])
-                            release["mac"]["check"] += "Fail "
-                        else:
-                            release["mac"]["check"] += "PASSED "
-
-                    else:
-                        # print(asset["browser_download_url"])
-                        release["mac"]["url"] = asset["browser_download_url"]
-                        if not check_url(asset["browser_download_url"]):
-                            errors.append(asset["browser_download_url"])
-                            release["mac"]["check"] = "Fail"
-                        else:
-                            release["mac"]["check"] = "PASSED"
-
-                    # print(asset["browser_download_url"])
-                    # release["mac"]["url"] = asset["browser_download_url"]
-                # else:
-                #     errors.append(asset["name"] + ": unrecognized asset.")
-            result.append(release)
-            print(str(count) + ": version " + json_element["tag_name"] + " will be processed.")
+    response = requests.get(f"{url}?per_page=100")
+    releases = response.json()
+    for release in releases:
+        if "TypeDB" in release["name"]:
+            print("Version " + release["tag_name"] + " will be processed.")
+            result.append(get_release_data(release))
         else:
-            print(str(count) + ": version " + json_element["tag_name"] + " IGNORED: no TypeDB in the name field.")
+            print("Version " + release["tag_name"] + " IGNORED: no TypeDB in the name field.")
     return result
 
 
-def generate_table_contents(versions, hash=False, tags=False):
+def get_release_data(release):
+    """Process each release data and extract information."""
+    data = {
+        "version": release["tag_name"],
+        "release_notes": release["html_url"],
+        "assets": {"win": {}, "lin": {}, "mac": {}}
+    }
+    for asset in release["assets"]:
+        name = asset["name"].lower()
+        if "typedb-all-linux" in name:
+            key = "arm64" if "arm64" in name else "x86_64" if "x86_64" in name else "url"
+            data["assets"]["lin"][key], data["assets"]["lin"]["check"] = get_asset_data(asset)
+        elif "typedb-all-mac" in name:
+            key = "arm64" if "arm64" in name else "x86_64" if "x86_64" in name else "url"
+            data["assets"]["mac"][key], data["assets"]["mac"]["check"] = get_asset_data(asset)
+        elif "typedb-all-windows" in name:
+            data["assets"]["win"]["url"], data["assets"]["win"]["check"] = get_asset_data(asset)
+    return data
+
+
+def get_asset_data(asset):
+    """Extract data from asset and verify URL."""
+    global ERRORS
+    url = asset["browser_download_url"]
+    if check_url(url):
+        check = "PASSED"
+    else:
+        check = "Fail"
+        ERRORS.append(url)
+    return url, check
+
+
+def generate_table_contents(versions, tags=False):
+    """Generate the table contents in asciidoc syntax."""
     result = ""
     for version in versions:
-        result += '\n| ' + version["release_notes"] + '[' + version["version"] + ']' + '\n'
-        # result += 'a|[,bash]' + '\n'
-        # result += '----' + '\n'
-        # result += 'docker pull vaticle/typedb:' + version["version"] + '\n'
-        # result += '----' + '\n'
-
-        for os in ["mac", "lin", "win"]:
-            result += '|'
+        result += f"\n| {version['release_notes']}[{version['version']}]\n"
+        for os_key in ["mac", "lin", "win"]:
+            result += '| '
             if tags:
-                result += '\n' + '// tag::' + os + '[]' + '\n'
+                result += f"\n// tag::{os_key}[]\n"
+            assets = version["assets"][os_key]
+            if assets.get("url"):
+                result += f"{assets['url']}[x86_64]\n"
             else:
-                result += ' '
-            if version[os]["url"] != "":
-                result += version[os]["url"] + '[x86_64]' + '\n'
-            else:
-                if version[os]["url-x86"] != "":
-                    result += version[os]["url-x86"] + '[x86_64] /'
-                if version[os]["url-arm"] != "":
-                    result += ' ' + version[os]["url-arm"] + '[arm64]' + '\n'
-
+                result += f"{assets['x86_64']}[x86_64] /"
+                result += f" {assets['arm64']}[arm64]\n"
             if tags:
-                result += '// end::' + os + '[]' + '\n'
-            result += '// Check: ' + version[os]["check"] + '\n'
-            if hash:
-                result += '| ' + version[os]["hash"] + '[_SHA256_]' + '\n'
+                result += f"// end::{os_key}[]\n"
+            result += f"// Check: {assets.get('check', '')}\n"
     return result
 
 
 def write_file(file_name, content):
-    with open(file_name, "w") as f:
-        f.write(content)
+    """Write content to the specified file."""
+    with open(file_name, "w") as file:
+        file.write(content)
 
 
 def print_json(url):
@@ -167,17 +91,21 @@ def print_json(url):
     for json_element in json_data:
         print(json_element)
 
-# print_json(repo)
 
-versions = get_versions(repo)
-for error in errors:
-    print("Warning! The following error occurred while checking asset urls:", error)
+if __name__ == "__main__":
+    """Main workflow"""
+    versions = get_versions(API_URL)
 
-all_downloads = generate_table_contents(versions)
-# print(all_downloads)
-write_file(filename_all, all_downloads)
-print("\nFile", filename_all, "write complete!")
+    all_downloads = generate_table_contents(versions)
+    try:
+        write_file(FILENAME_ALL, all_downloads)
+        print("\nFile", FILENAME_ALL, "write complete!")
+    except IOError:
+        print("Error while writing file:", FILENAME_ALL)
 
-latest_downloads = generate_table_contents([versions[0]], tags=True)
-write_file(filename_latest, latest_downloads)
-print("\nFile", filename_latest, "write complete!")
+    latest_downloads = generate_table_contents([versions[0]], True)
+    try:
+        write_file(FILENAME_LATEST, latest_downloads)
+        print("\nFile", FILENAME_LATEST, "write complete!")
+    except IOError:
+        print("Error while writing file:", FILENAME_LATEST)
