+++ TypeDB/all-versions.adoc
@@ -1,49 +1,49 @@
 
 | https://github.com/vaticle/typedb/releases/tag/2.25.6[2.25.6]
 | https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-mac-x86_64-2.25.6.zip[x86_64] / https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-mac-arm64-2.25.6.zip[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-linux-x86_64-2.25.6.tar.gz[x86_64] / https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-linux-arm64-2.25.6.tar.gz[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-windows-x86_64-2.25.6.zip[x86_64]
 // Check: PASSED
 
 | https://github.com/vaticle/typedb/releases/tag/2.25.0[2.25.0]
 | https://github.com/vaticle/typedb/releases/download/2.25.0/typedb-all-mac-x86_64-2.25.0.zip[x86_64] / https://github.com/vaticle/typedb/releases/download/2.25.0/typedb-all-mac-arm64-2.25.0.zip[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.25.0/typedb-all-linux-x86_64-2.25.0.tar.gz[x86_64] / https://github.com/vaticle/typedb/releases/download/2.25.0/typedb-all-linux-arm64-2.25.0.tar.gz[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.25.0/typedb-all-windows-x86_64-2.25.0.zip[x86_64]
 // Check: PASSED
 
 | https://github.com/vaticle/typedb/releases/tag/2.24.17[2.24.17]
 | https://github.com/vaticle/typedb/releases/download/2.24.17/typedb-all-mac-x86_64-2.24.17.zip[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.17/typedb-all-mac-arm64-2.24.17.zip[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.17/typedb-all-linux-x86_64-2.24.17.tar.gz[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.17/typedb-all-linux-arm64-2.24.17.tar.gz[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.17/typedb-all-windows-x86_64-2.24.17.zip[x86_64]
 // Check: PASSED
 
 | https://github.com/vaticle/typedb/releases/tag/2.24.16[2.24.16]
 | https://github.com/vaticle/typedb/releases/download/2.24.16/typedb-all-mac-x86_64-2.24.16.zip[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.16/typedb-all-mac-arm64-2.24.16.zip[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.16/typedb-all-linux-x86_64-2.24.16.tar.gz[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.16/typedb-all-linux-arm64-2.24.16.tar.gz[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.16/typedb-all-windows-x86_64-2.24.16.zip[x86_64]
 // Check: PASSED
 
 | https://github.com/vaticle/typedb/releases/tag/2.24.14[2.24.14]
 | https://github.com/vaticle/typedb/releases/download/2.24.14/typedb-all-mac-x86_64-2.24.14.zip[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.14/typedb-all-mac-arm64-2.24.14.zip[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.14/typedb-all-linux-x86_64-2.24.14.tar.gz[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.14/typedb-all-linux-arm64-2.24.14.tar.gz[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.14/typedb-all-windows-x86_64-2.24.14.zip[x86_64]
 // Check: PASSED
 
 | https://github.com/vaticle/typedb/releases/tag/2.24.12[2.24.12]
 | https://github.com/vaticle/typedb/releases/download/2.24.12/typedb-all-mac-x86_64-2.24.12.zip[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.12/typedb-all-mac-arm64-2.24.12.zip[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.12/typedb-all-linux-x86_64-2.24.12.tar.gz[x86_64] / https://github.com/vaticle/typedb/releases/download/2.24.12/typedb-all-linux-arm64-2.24.12.tar.gz[arm64]
-// Check: PASSED PASSED 
+// Check: PASSED
 | https://github.com/vaticle/typedb/releases/download/2.24.12/typedb-all-windows-x86_64-2.24.12.zip[x86_64]
 // Check: PASSED
