+++ TypeDB/latest-version.adoc
@@ -1,16 +1,16 @@
 
 | https://github.com/vaticle/typedb/releases/tag/2.25.6[2.25.6]
-|
+| 
 // tag::mac[]
 https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-mac-x86_64-2.25.6.zip[x86_64] / https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-mac-arm64-2.25.6.zip[arm64]
 // end::mac[]
-// Check: PASSED PASSED 
-|
+// Check: PASSED
+| 
 // tag::lin[]
 https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-linux-x86_64-2.25.6.tar.gz[x86_64] / https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-linux-arm64-2.25.6.tar.gz[arm64]
 // end::lin[]
-// Check: PASSED PASSED 
-|
+// Check: PASSED
+| 
 // tag::win[]
 https://github.com/vaticle/typedb/releases/download/2.25.6/typedb-all-windows-x86_64-2.25.6.zip[x86_64]
 // end::win[]
